Settings changes and scafolding for key translation routine.
 - Added KeyMapper class (WiP).
 - Default C++ idiom is now C++14

+++ ios/src/conflux_tests/MockSocket.m
@@ -58,7 +58,7 @@
 }
 
 -(void)recordWithHeader:(const UInt8 *)bytes bytes:(int)howMany {
-    UInt8 header[] = {0x00, 0x00, 0x00, howMany };
+    UInt8 header[] = {0x00, 0x00, 0x00, static_cast<UInt8>(howMany) };
     [self record:header bytes:sizeof(header)];
     [self record:bytes bytes:howMany];
 }
